Implemented index.py
made index.py actually do something, all it does is display an INT currently but its more then what it did.

Moved all (.css) files from CSS folder to new folder called static to allow for flask implementation, CSS folder left unchanged for backwards compatibility. Files will eventually need for their ref to be updated for flask implementation

currently index.html is the only file that runs with flask currently.

Should be noted that the "template" folder has been changed to look in the root dir of the project. This will need to be reimplemented if the file is modified