Fixed main page links
Pages now link together however this will only work when running the website from the main.py, this will be a common theme from now on so it is probably best the website will be operated exclusively from this.

Individual Articles do not currently have any links to them, they will be needed to be transferred to a standardized format with the data being stored in a database and accessed as required

Book tickets moved from 'events' folder to the root folder to facilitate the link movement easier.

Current song: Whatever It Takes - Imagine Dragons