Made Comments Visable
Made comments show on the related article, still no way other than actaully editing the database to add them but we are working on it.

The CSS for this page is being a real head-ache, probably because its the only page that hassent been fully redone, High probablity that this will be redone in the near future.

Current Song: Without You - Avicii